# api/openapi.yaml
openapi: 3.1.0
info:
  title: BotHub API
  version: 0.1.0
  description: |
    Contract between the Go dashboard and the PHP API.

    Rules:
    - The API is the only way to change configuration. The dashboard never touches the DB.
    - Every route except `/api/health`, `GET /api/v1/setup`, `POST /api/v1/setup` and
      `POST /api/v1/auth/login` needs a valid session. One central middleware checks this.
    - Every changing request (POST, PUT, PATCH, DELETE) needs the header `X-CSRF-Token`.
    - Errors never contain UI text. They contain an i18n key that the dashboard translates.
    - Bot tokens are write-only. The API never returns them.
    - Actions that need Discord return `202` with a job. Poll `GET /api/v1/jobs/{jobId}`.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /api/v1/bots/{botId}/modules:
    get:
      operationId: listBotModules
      tags: [guilds]
      summary: Module status of a bot (applies to all its guilds)
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/ModuleState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/modules/{moduleKey}:
    put:
      operationId: setBotModule
      tags: [guilds]
      summary: Enable or disable a module for a bot (all its guilds)
      parameters:
        - $ref: "#/components/parameters/BotId"
        - $ref: "#/components/parameters/ModuleKey"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled: { type: boolean }
      responses:
        "200":
          description: Module state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModuleState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/bots/{botId}/plugins:
    get:
      operationId: listBotPlugins
      tags: [plugins]
      summary: Plugins installed for one bot
      parameters:
        - $ref: "#/components/parameters/BotId"
      responses:
        "200":
          description: Installed plugins
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/InstalledPlugin" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: installBotPlugin
      tags: [plugins]
      summary: Install a plugin for one bot. The user must grant every requested permission.
      parameters:
        - $ref: "#/components/parameters/BotId"
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pluginId, version, grantedPermissions]
              properties:
                pluginId: { type: string }
                version: { $ref: "#/components/schemas/PluginVersion" }
                grantedPermissions:
                  type: array
                  items: { $ref: "#/components/schemas/PluginPermission" }
      responses:
        "202": { $ref: "#/components/responses/JobAccepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: bothub_session
...
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Token from `GET /api/v1/auth/me`
      schema: { type: string }
    BotId:
      name: botId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
    ModuleKey:
      name: moduleKey
      in: path
      required: true
      schema: { $ref: "#/components/schemas/ModuleKey" }
...
  responses:
    Unauthorized:
      description: No valid session (`error.auth.required`, `error.auth.invalid_credentials`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: CSRF token missing or wrong (`error.csrf.invalid`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (`error.<resource>.not_found`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State does not allow this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    JobAccepted:
      description: Job queued for the bot
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Job" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [key]
          properties:
            key:
              type: string
              description: i18n key, e.g. `error.bot.not_found`
              examples: [error.bot.not_found]
            params:
              type: object
              description: Values for placeholders in the translated text
              additionalProperties: true
            fields:
              type: object
              description: Per-field i18n keys for validation errors
              additionalProperties: { type: string }
...
    ModuleKey:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,63}$"
      examples: [economy, leveling, ticket]
...
    ModuleState:
      type: object
      required: [key, enabled]
      properties:
        key: { $ref: "#/components/schemas/ModuleKey" }
        enabled: { type: boolean }
...
    PluginVersion:
      type: string
      description: "BotHub scheme: feature X.0.0, bugfix 0.X.0, patch 0.0.X"
      pattern: "^[0-9]+\\.[0-9]+\\.[0-9]+$"
    PluginPermission:
      type: string
      enum:
        - discord.read_messages
        - discord.send_messages
        - discord.manage_messages
        - discord.manage_roles
        - discord.manage_channels
        - discord.moderate_members
        - discord.voice
        - db.own_tables
        - http.outbound
        - builder.nodes
        - dashboard.pages
...
    InstalledPlugin:
      type: object
      required: [id, name, version, enabled, grantedPermissions]
      properties:
        id: { type: string }
        name: { type: string }
        description:
          type: string
          description: Plugin-provided text in the plugin's default language
        icon: { type: string, description: "One emoji" }
        beta: { type: boolean }
        version: { $ref: "#/components/schemas/PluginVersion" }
        enabled: { type: boolean }
        grantedPermissions:
          type: array
          items: { $ref: "#/components/schemas/PluginPermission" }
        updateAvailable: { type: [string, "null"] }
...
    Job:
      type: object
      required: [id, type, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [bot.start, bot.stop, bot.restart, guild.leave, plugin.install, plugin.update, plugin.uninstall]
        status:
          type: string
          enum: [queued, running, done, failed]
        errorKey: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }